Try to eliminate render-arrows-before-ready
Original problem was font loading after the table rendered; I fixed that by using the WebFont API to delay setting up the table.  Now it appears to be rendering before the CSS is loaded?  Put scripts under style links (Not likely to help, but shouldn't hurt) and adding a tiny delay to the browser can give things a moment to shake out.