Adds options object and `ready` callback to ctor
This PR adds a simple options object with one parameter: ready. This parameter is a function which takes two arguments: error and id. error is null or undefined if HRMViewer is ready, otherwise it contains the error message. The id argument echos back the DOM ID of the element to differentiate code viewers.